Fixed the max digest loop error on the resume page
The resume page was endlessly running the digest loop (or at least too
many times to keep Angular happy.) The issue was that I had employed a
custom sorting function inside an ngRepeat directive, which was running
each and every time a new item was created in the list.

To fix it, I just employed the built-in orderBy filter, and extended the
asDate filter to turn 'Present' strings into the current date. I may go
back and update the asDate filter to turn 'Present' into the current day
at midnight, so that two items that both end on present are sorted
correctly, and not based on which one was parsed first. The effect is
probably minor, though.